Environment discrete -> continuous

+++ src/agents/AgentPeople.java
@@ -5,11 +5,11 @@ public class AgentPeople extends AgentType
 
 	private static final long serialVersionUID = 1L;
 
-	public AgentPeople(int x, int y) {
+	public AgentPeople(double x, double y) {
 		super(x, y);
 	}
 
-	public AgentPeople(int x, int y, double velocity, double radius, double angle) {
+	public AgentPeople(double x, double y, double velocity, double radius, double angle) {
 		super(x, y, velocity, radius, angle);
 	}
